test(platform-deno): colocate Redis credential coverage
